Fix rendering blocks outside of the maze

+++ src/polyomino.cpp
@@ -236,7 +236,7 @@ void Polyomino::render(int y_scroll) {
 
   banked_lambda(GET_BANK(polyominos), [this, y_scroll]() {
     definition->render(board.origin_x + (u8)(column << 4),
-                       (board.origin_y - y_scroll + (u8)(row << 4)));
+                       (board.origin_y - y_scroll + (row << 4)));
   });
 }
